backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
beanie>=1.26.0,<2.0.0
bcrypt>=4.2.0
PyJWT>=2.9.0
aiofiles>=24.1.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
email-validator>=2.2.0
httpx>=0.28.0
Pillow>=10.0.0
pillow-heif>=0.16.0
markdown>=3.5
